+++ test/parkour/io/avro_test.clj
@@ -5,8 +5,9 @@
             [clojure.core.reducers :as r]
             [abracad.avro :as avro]
             [parkour (conf :as conf) (fs :as fs) (wrapper :as w)
-                     (mapreduce :as mr)]
-            [parkour.io (dsink :as dsink) (avro :as mra)]))
+                     (mapreduce :as mr) (graph :as pg)]
+            [parkour.io (dsink :as dsink) (avro :as mra)]
+            [parkour.test-helpers :as th]))
 
 (deftest test-input
   (let [records ["foo" "bar" "baz" "quux"], schema (avro/parse-schema :string)
@@ -58,3 +59,28 @@
     (let [f (->> p fs/path-list (remove fs/hidden?) first io/file)]
       (is (= records (->> p (mra/dseq [:default]) w/unwrap (r/map first)
                           (into [])))))))
+
+(defn ->keys
+  [coll] (->> coll mr/keyvals (mr/sink-as :keys)))
+
+(defn kkg->kvg
+  [coll]
+  (->> (mr/keykeygroups coll)
+       (r/map (fn [[[k] ks]] [k (mapv #(nth % 1) ks)]))
+       (mr/sink-as :keyvals)))
+
+(deftest test-grouping
+  (let [inpath (doto (fs/path "tmp/input") fs/path-delete)
+        dseq (dsink/with-dseq (mra/dsink [:string :long] inpath)
+               [["a" 1] ["a" 2] ["b" 3] ["c" 4] ["c" 5]])
+        outpath (doto (fs/path "tmp/output") fs/path-delete)
+        dsink (mra/dsink [:string {:type "array", :items :long}] outpath)
+        schema (avro/tuple-schema [:string :long])
+        schema+g (avro/grouping-schema 1 schema)
+        [result] (-> (pg/input dseq)
+                     (pg/map #'->keys)
+                     (pg/partition (mra/shuffle schema nil schema+g))
+                     (pg/reduce #'kkg->kvg)
+                     (pg/output dsink)
+                     (pg/execute (th/config) "avro-grouping"))]
+    (is (= {"a" [1 2], "b" [3], "c", [4 5]} (into {} result)))))